remove all semver tests and just use the range test
